+++ demo/src/pages/HomePage.tsx
@@ -1,7 +1,8 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { AppRoute } from '../Routes';
 import { makeStyles, Theme, Box } from '@material-ui/core';
 import { Assessment } from '../common';
+import { getFromLocalStorage, saveToLocalStorageWithExpiration } from '../utils/localStorage';
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
@@ -300,22 +301,45 @@ const useStyles = makeStyles((theme: Theme) => ({
 const Dashboard = ({ history }) => {
   const classes = useStyles();
 
+  const serverEndpoint = getFromLocalStorage('serverEndpoint') ?? 'production';
+  const [selectedOption, setSelectedOption] = useState(serverEndpoint);
+
+  const handleOptionChange = (event) => {
+    setSelectedOption(event.target.value);
+    saveToLocalStorageWithExpiration('serverEndpoint', event.target.value)
+  };
+
   const menuItems = [
     {
-      name: `Squats`,
-      assessment_name: Assessment.SQUATS,
-      backgroundImage: `url('/Xtra Cards/strength.jpg')`,
+      name: 'Half Squat',
+      assessment_name: Assessment.HALF_SQUAT,
+      backgroundImage: `url('/Xtra Cards/sa.jpg')`,
     },
     {
-      name: `Range of Motion`,
-      assessment_name: Assessment.RANGE_OF_MOTION,
+      name: 'Banded Diagonal',
+      assessment_name: Assessment.BANDED_ALTERNATING_DIAGNOLS,
+      backgroundImage: `url('/Xtra Cards/2.jpg')`,
+    },
+    {
+      name: 'Sit Wall',
+      assessment_name: Assessment.SIT_WALL,
       backgroundImage: `url('/Xtra Cards/5.jpg')`,
     },
-    // {
-    //   name: `Side Flamingo`,
-    //   assessment_name: Assessment.SIDE_FLAMINGO,
-    //   backgroundImage: `url('/Xtra Cards/6.jpg')`,
-    // },
+    {
+      name: 'Push Ups',
+      assessment_name: Assessment.PUSH_UPS,
+      backgroundImage: `url('/Xtra Cards/6.jpg')`,
+    },
+    {
+      name: 'V Sit and Reach',
+      assessment_name: Assessment.SIT_AND_REACH_T2,
+      backgroundImage: `url('/Xtra Cards/cf.jpg')`,
+    },
+    {
+      name: 'Sit Ups',
+      assessment_name: Assessment.SIT_UPS_T2,
+      backgroundImage: `url('/Xtra Cards/strength.jpg')`,
+    }
   ];
 
   return (
@@ -327,7 +351,8 @@ const Dashboard = ({ history }) => {
             display: 'flex',
             flexDirection: 'row',
             marginTop: '30px',
-            justifyContent: 'space-between',
+            justifyContent: 'flex-start',
+            flexWrap: 'wrap'
           }}
         >
           {menuItems.map((item, indx) => (
@@ -349,8 +374,8 @@ const Dashboard = ({ history }) => {
                 className={classes.cardButton}
                 onClick={(e) => {
                   e.preventDefault();
-
                   history.push(AppRoute.Workout, {
+                    selectedOption,
                     assessment_name: item.assessment_name,
                   });
                 }}
@@ -359,6 +384,29 @@ const Dashboard = ({ history }) => {
               </div>
             </div>
           ))}
+        </div>
+        <div style={{
+          
+          position: 'absolute',/* Position the div absolutely within the body */
+          bottom: '20px',/* Distance from the bottom of the body */
+          right: '20px', /* Distance from the right of the body */
+          padding: '10px',/* Add some padding for content spacing */
+          // border: '1px solid #ccc',/* Add a border for visibility */
+          fontSize: '20px',
+          fontWeight: 'normal',
+        }}>
+          {/* <label>Choose an Environment:   </label> */}
+          <select  onChange={handleOptionChange} style={{
+            fontWeight: 'bold',
+            fontSize: '20px',
+            backgroundColor: 'rgb(35 226 245)',
+          }}
+          defaultValue={serverEndpoint}
+          >
+            <option value="production">Production</option>
+            <option value="staging">Staging</option>
+            <option value="local">Local</option>
+          </select>
         </div>
       </div>
     </Box>