+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # @xtravision/xtravision-react- Changelog
 
+### v2.4.0: (Thu Jul 27, 2023)
+- (sdk/update) Add provision to change url: prod/staging/local
+- (DemoApp/update) Add provision to change url: prod/staging/local
+
 ### v2.3.2: (May 15, 2023)
 - (sdk/hot-fix) Added XtraVisionEventEmitter for internal API